ModNet Errors Parse

+++ GameLauncher/Program.cs
@@ -30,32 +30,38 @@ namespace GameLauncher {
                 MessageBox.Show("There's no internet connection, launcher might crash");
             }
 
+            IniFile _settingFile = new IniFile("Settings.ini");
+
             //Windows 7 Fix
-            String _OS = (string)Registry.LocalMachine.OpenSubKey("Software\\Microsoft\\Windows NT\\CurrentVersion").GetValue("productName");
-            if(_OS.Contains("Windows 7")) {
-                if (Self.getInstalledHotFix("KB3125574") == false || Self.getInstalledHotFix("KB3125574") == false) {
-                    String messageBoxPopupKB = String.Empty;
-                    messageBoxPopupKB  = "Hey Windows 7 User, in order to play on this server, we need to make additional tweaks to your system.\n";
-                    messageBoxPopupKB += "We must make sure you have those Windows Update packages installed:\n\n";
+            if (_settingFile.Read("PatchesApplied") != "1") { 
+                String _OS = (string)Registry.LocalMachine.OpenSubKey("Software\\Microsoft\\Windows NT\\CurrentVersion").GetValue("productName");
+                if(_OS.Contains("Windows 7")) {
+                    if (Self.getInstalledHotFix("KB3125574") == false || Self.getInstalledHotFix("KB3125574") == false) {
+                        String messageBoxPopupKB = String.Empty;
+                        messageBoxPopupKB  = "Hey Windows 7 User, in order to play on this server, we need to make additional tweaks to your system.\n";
+                        messageBoxPopupKB += "We must make sure you have those Windows Update packages installed:\n\n";
 
-                    if (Self.getInstalledHotFix("KB3020369") == false) messageBoxPopupKB += "- Update KB3020369\n";
-                    if (Self.getInstalledHotFix("KB3125574") == false) messageBoxPopupKB += "- Update KB3125574\n";
+                        if (Self.getInstalledHotFix("KB3020369") == false) messageBoxPopupKB += "- Update KB3020369\n";
+                        if (Self.getInstalledHotFix("KB3125574") == false) messageBoxPopupKB += "- Update KB3125574\n";
 
-                    messageBoxPopupKB += "\nAditionally, we must add a value to the registry:\n";
+                        messageBoxPopupKB += "\nAditionally, we must add a value to the registry:\n";
 
-                    messageBoxPopupKB += "- HKLM/SYSTEM/CurrentControlSet/Control/SecurityProviders\n/SCHANNEL/Protocols/TLS 1.2/Client\n";
-                    messageBoxPopupKB += "- Value: DisabledByDefault -> 0\n\n";
+                        messageBoxPopupKB += "- HKLM/SYSTEM/CurrentControlSet/Control/SecurityProviders\n/SCHANNEL/Protocols/TLS 1.2/Client\n";
+                        messageBoxPopupKB += "- Value: DisabledByDefault -> 0\n\n";
 
-                    messageBoxPopupKB += "Would you like to add those values?";
-                    DialogResult replyPatchWin7 = MessageBox.Show(null, messageBoxPopupKB, "GameLauncherReborn", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                        messageBoxPopupKB += "Would you like to add those values?";
+                        DialogResult replyPatchWin7 = MessageBox.Show(null, messageBoxPopupKB, "GameLauncherReborn", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
 
-                    if(replyPatchWin7 == DialogResult.Yes) {
-                        RegistryKey key = Registry.LocalMachine.CreateSubKey(@"SYSTEM\CurrentControlSet\Control\SecurityProviders\SCHANNEL\Protocols\TLS 1.2\Client");
-                        key.SetValue("DisabledByDefault", 0x0);
+                        if(replyPatchWin7 == DialogResult.Yes) {
+                            RegistryKey key = Registry.LocalMachine.CreateSubKey(@"SYSTEM\CurrentControlSet\Control\SecurityProviders\SCHANNEL\Protocols\TLS 1.2\Client");
+                            key.SetValue("DisabledByDefault", 0x0);
 
-                        MessageBox.Show(null, "Registry option set, Remember that the following patch might work after a system reboot", "GameLauncherReborn", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    } else {
-                        MessageBox.Show(null, "Roger that, There will be some issues connecting to the servers.", "GameLauncherReborn", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                            _settingFile.Write("PatchesApplied", "1");
+
+                            MessageBox.Show(null, "Registry option set, Remember that the following patch might work after a system reboot", "GameLauncherReborn", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        } else {
+                            MessageBox.Show(null, "Roger that, There will be some issues connecting to the servers.", "GameLauncherReborn", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        }
                     }
                 }
             }
@@ -65,8 +71,6 @@ namespace GameLauncher {
             if (!Self.hasWriteAccessToFolder(Path.GetDirectoryName(Application.ExecutablePath))) {
                 MessageBox.Show("This application requires admin priviledge");
             }
-
-            IniFile _settingFile = new IniFile("Settings.ini");
 
             if (DetectLinux.LinuxDetected()) {
                 if (!_settingFile.KeyExists("InstallationDirectory")) {
